+++ packages/shared/src/components/profile/PostsSection.tsx
@@ -23,7 +23,8 @@ import EyeIcon from '../icons/Eye';
 import { largeNumberFormat } from '../../lib/numberFormat';
 import UpvoteIcon from '../icons/Upvote';
 import CommentIcon from '../icons/Discuss';
-import { AUTHOR_FEED_QUERY, FeedData, Post } from '../../graphql/posts';
+import { FeedData, Post } from '../../graphql/posts';
+import { AUTHOR_FEED_QUERY } from '../../graphql/feed';
 import { graphqlUrl } from '../../lib/config';
 import { LazyImage } from '../LazyImage';
 import { UserProfile } from '../../lib/user';
